Added test on cp module file. Close #31

+++ check-build.sh
@@ -61,7 +61,9 @@ MODULE_FILE
 ) > modules/${VERSION}
 mkdir -p ${LIBRARIES_MODULES}/${NAME}
 
-cp modules/${VERSION} ${LIBRARIES_MODULES}/${NAME}
+if [! -e ${LIBRARIES_MODULES}/${NAME}/${VERSION}] ; then
+  cp modules/${VERSION} ${LIBRARIES_MODULES}/${NAME}
+fi
 
 echo "[check-build.sh] - Checking REPAST module"
 module add $NAME/$VERSION
